Renamed OpenSimScene to SimpleModelScene
